requirements.txt
streamlit==1.19.0
openai
python-dotenv
moviepy